Small update
Publishing on blast now works
Added a few more scripts
Fixed the chain id bug

+++ src/Echo.sol
@@ -15,12 +15,12 @@ contract Echo is ReentrancyGuard, Admin, Banker, Escrow, EchoState {
     event OfferAccepted(bytes32 indexed offerId);
     event OfferCanceled(bytes32 indexed offerId);
     event OfferExecuted(bytes32 indexed offerId);
-    // @dev For future use...
 
-    uint16 private immutable CHAIN_ID;
+    // @dev For future use...
+    uint256 private immutable CHAIN_ID;
 
     constructor(address owner) Admin(owner) {
-        CHAIN_ID = uint16(block.chainid);
+        CHAIN_ID = block.chainid;
     }
 
     /**
